Added more logs in packets.

+++ src/packet/ListenDatagramUDP.cpp
@@ -57,7 +57,7 @@ vector<uint8_t> ListenDatagramUDP::receivePacket()
 
     cout << "Frame: ";
     for ( auto a : frame) {
-        cout << a ;
+        cout << static_cast<int>(a) ;
     }
     cout << endl;
